Descripción breve de los cambios realizados

+++ Almacen_backend/src/main/java/com/QF/Almacen_backend/Repositorios/RepsitorioProductos.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepsitorioProductos extends JpaRepository<Productos,Long> {
+public interface RepsitorioProductos extends JpaRepository<Productos,Integer> {
 }
